handle delete file when delete product

+++ controllers/admin.js
@@ -1,4 +1,6 @@
 const { validationResult } = require("express-validator");
+const fileHelper = require("../util/file");
+
 const Product = require("../models/product");
 
 exports.getAddProduct = (req, res, next) => {
@@ -14,16 +16,33 @@ exports.getAddProduct = (req, res, next) => {
 
 exports.postAddProduct = (req, res, next) => {
 	const title = req.body.title;
-	const imageUrl = req.body.imageUrl;
+	const image = req.file;
 	const price = req.body.price;
 	const description = req.body.description;
+
+	if (!image) {
+		return res.status(422).render("admin/edit-product", {
+			pageTitle: "Add Product",
+			path: "/admin/add-product",
+			editing: false,
+			hasError: true,
+			product: {
+				title: title,
+				price: price,
+				description: description,
+			},
+			errorMessage: "Attached file is not an image.",
+			validationErrors: [],
+		});
+	}
+
 	const errors = validationResult(req);
 
 	if (!errors.isEmpty()) {
 		console.log(errors.array());
 		return res.status(422).render("admin/edit-product", {
 			pageTitle: "Add Product",
-			path: "/admin/edit-product",
+			path: "/admin/add-product",
 			editing: false,
 			hasError: true,
 			product: {
@@ -37,6 +56,7 @@ exports.postAddProduct = (req, res, next) => {
 		});
 	}
 
+	const imageUrl = image.path;
 	const product = new Product({
 		title: title,
 		price: price,
@@ -47,11 +67,12 @@ exports.postAddProduct = (req, res, next) => {
 	product
 		.save()
 		.then((result) => {
-			console.log("Created Product");
 			res.redirect("/admin/products");
 		})
 		.catch((err) => {
-			console.log(err);
+			const error = new Error(err);
+			error.httpStatusCode = 500;
+			return next(error);
 		});
 };
 
@@ -76,14 +97,18 @@ exports.getEditProduct = (req, res, next) => {
 				validationErrors: [],
 			});
 		})
-		.catch((err) => console.log(err));
+		.catch((err) => {
+			const error = new Error(err);
+			error.httpStatusCode = 500;
+			return next(error);
+		});
 };
 
 exports.postEditProduct = (req, res, next) => {
 	const prodId = req.body.productId;
 	const updatedTitle = req.body.title;
 	const updatedPrice = req.body.price;
-	const updatedImageUrl = req.body.imageUrl;
+	const image = req.file;
 	const updatedDesc = req.body.description;
 
 	const errors = validationResult(req);
@@ -96,7 +121,6 @@ exports.postEditProduct = (req, res, next) => {
 			hasError: true,
 			product: {
 				title: updatedTitle,
-				imageUrl: updatedImageUrl,
 				price: updatedPrice,
 				description: updatedDesc,
 				_id: prodId,
@@ -118,15 +142,20 @@ exports.postEditProduct = (req, res, next) => {
 			product.title = updatedTitle;
 			product.price = updatedPrice;
 			product.description = updatedDesc;
-			product.imageUrl = updatedImageUrl;
-
+			if (image) {
+				fileHelper.deleteFile(product.imageUrl);
+				product.imageUrl = image.path;
+			}
 			return product.save().then((result) => {
 				console.log("UPDATED PRODUCT!");
 				res.redirect("/admin/products");
 			});
 		})
-
-		.catch((err) => console.log(err));
+		.catch((err) => {
+			const error = new Error(err);
+			error.httpStatusCode = 500;
+			return next(error);
+		});
 };
 
 exports.getProducts = (req, res, next) => {
@@ -140,15 +169,31 @@ exports.getProducts = (req, res, next) => {
 				path: "/admin/products",
 			});
 		})
-		.catch((err) => console.log(err));
+		.catch((err) => {
+			const error = new Error(err);
+			error.httpStatusCode = 500;
+			return next(error);
+		});
 };
 
 exports.postDeleteProduct = (req, res, next) => {
 	const prodId = req.body.productId;
-	Product.deleteOne({ _id: prodId, userId: req.user._id })
+
+	Product.findById(prodId)
+		.then((product) => {
+			if (!product) {
+				return next(new Error("Product not found."));
+			}
+			fileHelper.deleteFile(product.imageUrl);
+			return Product.deleteOne({ _id: prodId, userId: req.user._id });
+		})
 		.then(() => {
 			console.log("DESTROYED PRODUCT");
 			res.redirect("/admin/products");
 		})
-		.catch((err) => console.log(err));
+		.catch((err) => {
+			const error = new Error(err);
+			error.httpStatusCode = 500;
+			return next(error);
+		});
 };
